refactor: migrate high risks
ref #2318
