+++ visualization/app/codeCharta/ui/artificialIntelligence/artificialIntelligence.component.html
@@ -67,68 +67,4 @@
 			</ul>
 		</md-menu-content>
 	</md-menu>
-
-	<md-menu>
-		<md-button class="high-risk-profile-button" ng-click="$mdMenu.open($event)" title="Open High Risk Profile">
-			<i class="fa fa-bar-chart"></i>
-		</md-button>
-
-		<md-menu-content class="ai-drop-down summary">
-			<md-menu-item class="title" ng-show="$ctrl.viewModel.analyzedProgrammingLanguage.length">
-				<p>Risk Profile</p>
-				<p>Calculations are always based on mcc (cyclomatic complexity)</p>
-			</md-menu-item>
-
-			<md-menu-item class="suspicious-metrics-summary">
-				<ul ng-show="$ctrl.viewModel.analyzedProgrammingLanguage.length">
-					<li>{{ $ctrl.viewModel.riskProfile.lowRisk }}% of overall code is in files with low complexity</li>
-					<li>{{ $ctrl.viewModel.riskProfile.moderateRisk }}% of overall code is in files with moderate complexity</li>
-					<li>{{ $ctrl.viewModel.riskProfile.highRisk }}% of overall code is in files with high complexity</li>
-					<li>{{ $ctrl.viewModel.riskProfile.veryHighRisk }}% of overall code is in files with extreme complexity</li>
-				</ul>
-				<span ng-show="$ctrl.viewModel.analyzedProgrammingLanguage === undefined">
-					Could not calculate Risk Profile. Metrics rloc and mcc not available.
-				</span>
-			</md-menu-item>
-
-			<md-menu-item class="risk-profiles" ng-show="$ctrl.viewModel.analyzedProgrammingLanguage.length">
-				<div
-					ng-if="$ctrl.viewModel.riskProfile.lowRisk > 0"
-					title="Low Risk ({{ $ctrl.viewModel.riskProfile.lowRisk }}% of code)"
-					class="risk-profile low-risk"
-					style="width:{{ $ctrl.viewModel.riskProfile.lowRisk }}%;
-                 {{ $ctrl.viewModel.riskProfile.lowRisk < 5 ? 'color:rgba(0,0,0,0)' : '' }}"
-				>
-					{{ $ctrl.viewModel.riskProfile.lowRisk }}%
-				</div>
-				<div
-					ng-if="$ctrl.viewModel.riskProfile.moderateRisk > 0"
-					title="Moderate Risk ({{ $ctrl.viewModel.riskProfile.moderateRisk }}% of code)"
-					class="risk-profile moderate-risk"
-					style="width:{{ $ctrl.viewModel.riskProfile.moderateRisk }}%;
-                {{ $ctrl.viewModel.riskProfile.moderateRisk < 5 ? 'color:rgba(0,0,0,0)' : '' }}"
-				>
-					{{ $ctrl.viewModel.riskProfile.moderateRisk }}%
-				</div>
-				<div
-					ng-if="$ctrl.viewModel.riskProfile.highRisk > 0"
-					title="High Risk ({{ $ctrl.viewModel.riskProfile.highRisk }}% of code)"
-					class="risk-profile high-risk"
-					style="width:{{ $ctrl.viewModel.riskProfile.highRisk }}%;
-                {{ $ctrl.viewModel.riskProfile.highRisk < 5 ? 'color:rgba(0,0,0,0)' : '' }}"
-				>
-					{{ $ctrl.viewModel.riskProfile.highRisk }}%
-				</div>
-				<div
-					ng-if="$ctrl.viewModel.riskProfile.veryHighRisk > 0"
-					title="Very High Risk ({{ $ctrl.viewModel.riskProfile.veryHighRisk }}% of code)"
-					class="risk-profile very-high-risk"
-					style="width:{{ $ctrl.viewModel.riskProfile.veryHighRisk }}%;
-                {{ $ctrl.viewModel.riskProfile.veryHighRisk < 5 ? 'color:rgba(0,0,0,0)' : '' }}"
-				>
-					{{ $ctrl.viewModel.riskProfile.veryHighRisk }}%
-				</div>
-			</md-menu-item>
-		</md-menu-content>
-	</md-menu>
 </div>
