--- visualization/app/codeCharta/ui/artificialIntelligence/artificialIntelligence.module.ts
@@ -1,7 +1,0 @@
-import "../../state/state.module"
-import angular from "angular"
-import { artificialIntelligenceComponent } from "./artificialIntelligence.component"
-
-angular
-	.module("app.codeCharta.ui.artificialIntelligence", ["app.codeCharta.state"])
-	.component(artificialIntelligenceComponent.selector, artificialIntelligenceComponent)
