+++ visualization/app/codeCharta/ui/ribbonBar/artificialIntelligence/artificialIntelligence.component.html
@@ -1,2 +1,2 @@
-<cc-suspicious-metrics></cc-suspicious-metrics>
-<cc-high-risk-profile></cc-high-risk-profile>
+<cc-suspicious-metrics [data]="data$ | async"></cc-suspicious-metrics>
+<cc-high-risk-profile [data]="data$ | async"></cc-high-risk-profile>
