+++ visualization/app/codeCharta/ui/ribbonBar/artificialIntelligence/highRiskProfile/highRiskProfile.component.ts
@@ -1,7 +1,10 @@
-import { Component } from "@angular/core"
+import { Component, Input } from "@angular/core"
+import { ArtificialIntelligenceData } from "../selectors/artificialIntelligence.selector"
 
 @Component({
 	selector: "cc-high-risk-profile",
 	template: require("./highRiskProfile.component.html")
 })
-export class HighRiskProfileComponent {}
+export class HighRiskProfileComponent {
+	@Input() data: ArtificialIntelligenceData
+}
