+++ visualization/app/codeCharta/ui/ribbonBar/artificialIntelligence/suspiciousMetrics/suspiciousMetrics.component.ts
@@ -1,7 +1,10 @@
-import { Component } from "@angular/core"
+import { Component, Input } from "@angular/core"
+import { ArtificialIntelligenceData } from "../selectors/artificialIntelligence.selector"
 
 @Component({
 	selector: "cc-suspicious-metrics",
 	template: require("./suspiciousMetrics.component.html")
 })
-export class SuspiciousMetricComponent {}
+export class SuspiciousMetricComponent {
+	@Input() data: ArtificialIntelligenceData
+}
