+++ visualization/app/codeCharta/ui/ribbonBar/artificialIntelligence/util/getMetricThresholds.ts
@@ -1,4 +1,4 @@
-import { metricThresholdsByLanguage } from "../artificialIntelligence.metricThresholds"
+import { metricThresholdsByLanguage } from "./artificialIntelligence.metricThresholds"
 
 export function getAssociatedMetricThresholds(programmingLanguage: string) {
 	return programmingLanguage === "java" ? metricThresholdsByLanguage.java : metricThresholdsByLanguage.miscellaneous
