+++ visualization/app/codeCharta/ui/ribbonBar/artificialIntelligence/util/mainProgrammingLanguageHelper.ts
