+++ visualization/app/codeCharta/ui/ui.ts
@@ -1,7 +1,6 @@
 import angular from "angular"
 import { downgradeComponent } from "@angular/upgrade/static"
 
-import "./artificialIntelligence/artificialIntelligence.module"
 import "./screenshotButton/screenshotButton.module"
 import "./presentationModeButton/presentationModeButton.module"
 import "./codeMap/codeMap.module"
@@ -16,7 +15,6 @@ import { SliderComponent } from "./slider/slider.component"
 
 angular
 	.module("app.codeCharta.ui", [
-		"app.codeCharta.ui.artificialIntelligence",
 		"app.codeCharta.ui.presentationModeButton",
 		"app.codeCharta.ui.codeMap",
 		"app.codeCharta.ui.dialog",
